upgrade to SSDK16; springt boot 3; ignored modules; cursor & composition;
