+++ pom.xml
@@ -5,8 +5,8 @@
     <parent>
         <groupId>com.backbase.buildingblocks</groupId>
         <artifactId>backbase-parent</artifactId>
-        <version>15.2.4</version>
-        <relativePath />
+        <version>16.0.1</version>
+        <relativePath/>
     </parent>
 
     <groupId>com.backbase.stream</groupId>
@@ -23,7 +23,7 @@
         <module>stream-product</module>
         <module>stream-access-control</module>
         <module>stream-transactions</module>
-        <module>stream-cursor</module>
+        <!--        <module>stream-cursor</module>-->
         <module>stream-product-catalog</module>
         <module>stream-limits</module>
         <module>stream-legal-entity</module>
@@ -31,16 +31,16 @@
         <module>stream-payment-order</module>
         <module>stream-portfolio</module>
         <module>stream-contacts</module>
-        <module>stream-compositions</module>
-        <module>stream-loans</module>
-        <module>stream-audiences</module>
+        <!--        <module>stream-compositions</module>-->
+        <!--        <module>stream-loans</module>-->
+        <!--        <module>stream-audiences</module>-->
     </modules>
 
     <properties>
         <java.version>17</java.version>
-        <ssdk.version>15.2.4</ssdk.version>
-        <backbase-bom.version>2023.09.17-LTS</backbase-bom.version>
-        <boat-maven-plugin.version>0.16.6</boat-maven-plugin.version>
+        <ssdk.version>16.0.1</ssdk.version>
+        <backbase-bom.version>2023.12</backbase-bom.version>
+        <boat-maven-plugin.version>0.17.24</boat-maven-plugin.version>
         <spring-boot.repackage.skip>true</spring-boot.repackage.skip>
         <openapi-generator-maven-plugin.language>spring</openapi-generator-maven-plugin.language>
         <openapi-generator-maven-plugin.library>spring-boot</openapi-generator-maven-plugin.library>
