+++ stream-access-control/access-control-core/src/main/java/com/backbase/stream/mapper/LegalEntityMapper.java
@@ -29,7 +29,6 @@ public interface LegalEntityMapper {
 
     com.backbase.stream.legalentity.model.LegalEntityType map(LegalEntityType legalEntityType);
 
-    @Mapping(source = "type", target = "legalEntityType")
     LegalEntity toModel(LegalEntityCreateItem legalEntity);
 
     @Mapping(source = "id", target = "internalId")
