+++ stream-compositions/pom.xml
@@ -7,7 +7,7 @@
     <parent>
         <groupId>com.backbase.buildingblocks</groupId>
         <artifactId>service-sdk-starter-reactive</artifactId>
-        <version>15.2.4</version>
+        <version>16.0.1</version>
         <relativePath/>
     </parent>
 
@@ -16,14 +16,14 @@
     <version>4.0.0-SNAPSHOT</version>
 
     <properties>
-        <backbase-bom.version>2023.09.17-LTS</backbase-bom.version>
-        <boat-maven-plugin.version>0.16.6</boat-maven-plugin.version>
+        <backbase-bom.version>2023.12</backbase-bom.version>
+        <boat-maven-plugin.version>0.17.24</boat-maven-plugin.version>
         <testcontainers.version>1.12.5</testcontainers.version>
         <mock.server.version>5.11.2</mock.server.version>
         <skip.integration.tests>false</skip.integration.tests>
         <sonar.test.exclusions>**/test/**/*</sonar.test.exclusions>
-        <spring-cloud-kubernetes.version>2.1.3</spring-cloud-kubernetes.version>
-        <ssdk.version>15.2.4</ssdk.version>
+        <spring-cloud-kubernetes.version>3.0.4</spring-cloud-kubernetes.version>
+        <ssdk.version>16.0.1</ssdk.version>
         <jib.container.volumes>/tmp</jib.container.volumes>
     </properties>
