+++ stream-dbs-clients/src/main/java/com/backbase/stream/clients/config/CompositeApiClientConfig.java
@@ -1,8 +1,8 @@
 package com.backbase.stream.clients.config;
 
 import com.backbase.buildingblocks.webclient.client.ApiClientConfig;
+import jakarta.validation.constraints.Positive;
 import java.util.Optional;
-import javax.validation.constraints.Positive;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cloud.client.ServiceInstance;
