+++ stream-legal-entity/legal-entity-http/src/main/java/com/backbase/stream/controller/ServiceAgreementController.java
@@ -2,8 +2,11 @@ package com.backbase.stream.controller;
 
 import com.backbase.stream.UpdatedServiceAgreementSaga;
 import com.backbase.stream.UpdatedServiceAgreementTask;
-import com.backbase.stream.legalentity.api.ServiceAgreementApi;
+import com.backbase.stream.UpdatedServiceAgreementUnitOfWorkExecutor;
+import com.backbase.stream.legalentity.ServiceAgreementApi;
 import com.backbase.stream.legalentity.model.UpdatedServiceAgreement;
+import com.backbase.stream.legalentity.model.UpdatedServiceAgreementResponse;
+import java.time.Duration;
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.ResponseEntity;
@@ -15,9 +18,18 @@ import reactor.core.publisher.Mono;
 @RestController
 @AllArgsConstructor
 @Slf4j
-public class ServiceAgreementController implements ServiceAgreementApi {
+public class ServiceAgreementController extends BaseAsyncController implements ServiceAgreementApi {
 
     private final UpdatedServiceAgreementSaga updatedServiceAgreementService;
+    private final UpdatedServiceAgreementUnitOfWorkExecutor updatedServiceAgreementUnitOfWorkExecutor;
+
+    @Override
+    public Mono<ResponseEntity<UpdatedServiceAgreementResponse>> getUpdateServiceAgreementUnitOfWork(
+        String unitOfWorkId, ServerWebExchange exchange) {
+        return updatedServiceAgreementUnitOfWorkExecutor.retrieve(unitOfWorkId)
+            .map(unitOfWorkMapper::convertToUpdatedServiceAgreementResponse)
+            .map(ResponseEntity::ok);
+    }
 
     @Override
     public Mono<ResponseEntity<Flux<UpdatedServiceAgreement>>> updateServiceAgreement(
@@ -30,4 +42,17 @@ public class ServiceAgreementController implements ServiceAgreementApi {
 
         return Mono.just(ResponseEntity.ok(flux));
     }
+
+    @Override
+    public Mono<ResponseEntity<Flux<UpdatedServiceAgreementResponse>>> updateServiceAgreementAsync(
+        Flux<UpdatedServiceAgreement> updatedServiceAgreement,
+        ServerWebExchange exchange) {
+        Flux<UpdatedServiceAgreementResponse> map = updatedServiceAgreement.bufferTimeout(10, Duration.ofMillis(100))
+            .map(this::createServiceAgreementUnitOfWork)
+            .flatMap(updatedServiceAgreementUnitOfWorkExecutor::register)
+            .map(unitOfWorkMapper::convertToUpdatedServiceAgreementResponse);
+        return Mono.just(ResponseEntity.ok(map));
+    }
+
+
 }
