+++ stream-models/legal-entity-model/pom.xml
@@ -54,23 +54,18 @@
                 <executions>
                     <execution>
                         <id>cursor</id>
-                        <phase>generate-sources</phase>
                         <goals>
-                            <goal>generate</goal>
+                            <goal>generate-spring-boot-embedded</goal>
                         </goals>
+                        <phase>generate-sources</phase>
                         <configuration>
-                            <generatorName>spring</generatorName>
+                            <openapiNormalizer>REFACTOR_ALLOF_WITH_PROPERTIES_ONLY=true</openapiNormalizer>
                             <inputSpec>${raml2openapi-maven-plugin.input}</inputSpec>
-                            <generateApiTests>false</generateApiTests>
-                            <generateApis>false</generateApis>
-                            <generateSupportingFiles>false</generateSupportingFiles>
                             <output>${openapi-generator-maven-plugin.output}</output>
+                            <modelPackage>com.backbase.stream.legalentity.model</modelPackage>
+                            <apiPackage>com.backbase.stream.legalentity</apiPackage>
                             <configOptions>
-                                <library>spring-boot</library>
                                 <reactive>true</reactive>
-                                <java8>true</java8>
-                                <dateLibrary>java8</dateLibrary>
-                                <modelPackage>com.backbase.stream.legalentity.model</modelPackage>
                                 <useBeanValidation>false</useBeanValidation>
                                 <useReflectionEqualsHashCode>false</useReflectionEqualsHashCode>
                             </configOptions>
@@ -79,7 +74,6 @@
                             </typeMappings>
                         </configuration>
                     </execution>
-
                 </executions>
             </plugin>
