+++ stream-models/portfolio-model/pom.xml
@@ -57,23 +57,17 @@
                 <executions>
                     <execution>
                         <id>cursor</id>
-                        <phase>generate-sources</phase>
                         <goals>
-                            <goal>generate</goal>
+                            <goal>generate-spring-boot-embedded</goal>
                         </goals>
+                        <phase>generate-sources</phase>
                         <configuration>
-                            <generatorName>spring</generatorName>
                             <inputSpec>${openapi-maven-plugin.input}</inputSpec>
-                            <generateApiTests>false</generateApiTests>
-                            <generateApis>false</generateApis>
-                            <generateSupportingFiles>false</generateSupportingFiles>
                             <output>${openapi-generator-maven-plugin.output}</output>
+                            <modelPackage>com.backbase.stream.portfolio.model</modelPackage>
+                            <apiPackage>com.backbase.stream.portfolio</apiPackage>
                             <configOptions>
-                                <library>spring-boot</library>
                                 <reactive>true</reactive>
-                                <java8>true</java8>
-                                <dateLibrary>java8</dateLibrary>
-                                <modelPackage>com.backbase.stream.portfolio.model</modelPackage>
                                 <useBeanValidation>false</useBeanValidation>
                                 <useReflectionEqualsHashCode>false</useReflectionEqualsHashCode>
                             </configOptions>
