+++ stream-portfolio/portfolio-http/pom.xml
@@ -33,7 +33,7 @@
         <dependency>
             <groupId>com.atlassian.oai</groupId>
             <artifactId>swagger-request-validator-core</artifactId>
-            <version>2.9.0</version>
+            <version>2.39.0</version>
         </dependency>
 
         <dependency>
@@ -65,54 +65,6 @@
                 <configuration>
                     <skip>true</skip>
                 </configuration>
-            </plugin>
-            <plugin>
-                <groupId>com.backbase.oss</groupId>
-                <artifactId>boat-maven-plugin</artifactId>
-                <executions>
-                    <execution>
-                        <id>generate</id>
-                        <phase>generate-sources</phase>
-                        <goals>
-                            <goal>generate</goal>
-                        </goals>
-                        <configuration>
-                            <generatorName>spring</generatorName>
-                            <inputSpec>${openapi-maven-plugin.input}</inputSpec>
-                            <generateApiTests>false</generateApiTests>
-                            <generateApis>true</generateApis>
-                            <generateModels>false</generateModels>
-                            <generateSupportingFiles>false</generateSupportingFiles>
-                            <output>${openapi-generator-maven-plugin.output}</output>
-                            <configOptions>
-                                <library>spring-boot</library>
-                                <reactive>true</reactive>
-                                <java8>true</java8>
-                                <dateLibrary>java8</dateLibrary>
-                                <interfaceOnly>true</interfaceOnly>
-                                <skipDefaultInterface>true</skipDefaultInterface>
-                                <modelPackage>${openapi-maven-plugin.base-package}.model</modelPackage>
-                                <apiPackage>${openapi-maven-plugin.base-package}.api</apiPackage>
-                                <useBeanValidation>false</useBeanValidation>
-                            </configOptions>
-                            <typeMappings>
-                                <typeMapping>Double=java.math.BigDecimal</typeMapping>
-                            </typeMappings>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>docs</id>
-                        <goals>
-                            <goal>generate</goal>
-                        </goals>
-                        <phase>site</phase>
-                        <configuration>
-                            <generatorName>html2</generatorName>
-                            <inputSpec>${openapi-maven-plugin.output}</inputSpec>
-                            <output>${project.build.directory}/site</output>
-                        </configuration>
-                    </execution>
-                </executions>
             </plugin>
 
             <plugin>
