+++ stream-product-catalog/product-catalog-task/src/main/java/com/backbase/stream/productcatalog/configuration/ProductCatalogConfigurationProperties.java
@@ -1,7 +1,7 @@
 package com.backbase.stream.productcatalog.configuration;
 
 import com.backbase.stream.productcatalog.model.ProductCatalog;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.validation.annotation.Validated;
