+++ stream-sdk/stream-parent/pom.xml
@@ -5,8 +5,8 @@
     <parent>
         <groupId>com.backbase.buildingblocks</groupId>
         <artifactId>backbase-parent</artifactId>
-        <version>15.2.4</version>
-        <relativePath />
+        <version>16.0.1</version>
+        <relativePath/>
     </parent>
 
     <groupId>com.backbase.stream</groupId>
@@ -22,10 +22,10 @@
     </organization>
 
     <licenses>
-        <license />
+        <license/>
     </licenses>
     <developers>
-        <developer />
+        <developer/>
     </developers>
 
     <modules>
@@ -38,7 +38,7 @@
 
     <properties>
         <java.version>17</java.version>
-        <ssdk.version>15.2.4</ssdk.version>
+        <ssdk.version>16.0.1</ssdk.version>
         <japicmp.skip>true</japicmp.skip>
     </properties>
