+++ stream-sdk/stream-parent/stream-worker/pom.xml
@@ -36,6 +36,7 @@
         <dependency>
             <groupId>org.springframework.cloud</groupId>
             <artifactId>spring-cloud-sleuth-api</artifactId>
+            <version>3.1.9</version>
         </dependency>
 
         <dependency>
