+++ stream-sdk/stream-starter-parents/pom.xml
@@ -20,7 +20,7 @@
     </organization>
 
     <licenses>
-        <license />
+        <license/>
     </licenses>
     <developers>
         <developer>
