+++ stream-sdk/stream-starter-parents/stream-http-starter-parent/pom.xml
@@ -6,7 +6,7 @@
     <parent>
         <groupId>com.backbase.buildingblocks</groupId>
         <artifactId>service-sdk-starter-reactive</artifactId>
-        <version>15.2.4</version>
+        <version>16.0.1</version>
         <relativePath/>
     </parent>
 
@@ -28,7 +28,7 @@
         <codegen.generated-sources-dir>${project.build.directory}/generated-sources</codegen.generated-sources-dir>
         <annotations.generated-sources-dir>${codegen.generated-sources-dir}/annotations</annotations.generated-sources-dir>
         <openapi-generator-maven-plugin.output>${codegen.generated-sources-dir}/openapi</openapi-generator-maven-plugin.output>
-        <spring-cloud-kubernetes.version>2.1.3</spring-cloud-kubernetes.version>
+        <spring-cloud-kubernetes.version>3.0.4</spring-cloud-kubernetes.version>
     </properties>
 
     <dependencyManagement>
@@ -73,7 +73,7 @@
                 <plugin>
                     <groupId>com.backbase.oss</groupId>
                     <artifactId>boat-maven-plugin</artifactId>
-                    <version>0.15.8</version>
+                    <version>0.17.24</version>
                 </plugin>
                 <plugin>
                     <groupId>org.codehaus.mojo</groupId>
