+++ stream-sdk/stream-starter-parents/stream-task-starter-parent/pom.xml
@@ -5,7 +5,7 @@
     <parent>
         <groupId>com.backbase.buildingblocks</groupId>
         <artifactId>service-sdk-starter-reactive</artifactId>
-        <version>15.2.4</version>
+        <version>16.0.1</version>
         <relativePath/>
     </parent>
 
@@ -24,7 +24,7 @@
         <docker.scs.kafka>provided</docker.scs.kafka>
         <docker.scs.azure>provided</docker.scs.azure>
 
-        <spring-cloud-kubernetes.version>2.1.3</spring-cloud-kubernetes.version>
+        <spring-cloud-kubernetes.version>3.0.4</spring-cloud-kubernetes.version>
     </properties>
 
     <dependencyManagement>
