+++ stream-transactions/transactions-core/pom.xml
@@ -37,7 +37,6 @@
             <version>${project.version}</version>
         </dependency>
 
-
     </dependencies>
 
 </project>
